feat: Chage Mode V1
