+++ users/urls.py
@@ -1,23 +1,16 @@
 from django.urls import path
 
-from users.views.base_views import (
-    ChangeModeAPI,
-    ChangePassword,
-    JWTLogIn,
-    LogIn,
-    LogOut,
-    PublicUser,
-)
+from users.views.base_views import ChangePassword, JWTLogIn, LogIn, LogOut, PublicUser
 from users.views.view_v0_social_login import GithubLogIn, KakaoLogIn
 from users.views.view_v1_signup import SignUp
-from users.views.view_v1_user import MyProfileAPI
+from users.views.view_v1_user import ChangeModeAPI, MyProfileAPI
 
 urlpatterns = [
-    path("v0/change-mode", ChangeModeAPI.as_view()),
     path("v0/change-password", ChangePassword.as_view()),
     path("v0/jwt-login", JWTLogIn.as_view()),
     path("v1/sign-up", SignUp.as_view()),
     path("v1/me", MyProfileAPI.as_view()),
+    path("v1/change-mode", ChangeModeAPI.as_view()),
     path("v0/log-in", LogIn.as_view()),
     path("v0/log-out", LogOut.as_view()),
     path("v0/github", GithubLogIn.as_view()),
