+++ users/views/base_views.py
@@ -4,13 +4,11 @@ from django.contrib.auth import authenticate, login, logout
 from rest_framework import status
 from rest_framework.exceptions import NotFound, ParseError
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from users.models import User
 from users.serializers import PrivateUserSerializer
-from users.services.service_v1_user import UserService
 
 
 class Me(APIView):
@@ -103,11 +101,3 @@ class JWTLogIn(APIView):
             )
             return Response({"token": token})
         return Response({"error": "wrong password"})
-
-
-class ChangeModeAPI(APIView):
-    def patch(self, request: Request):
-        service = UserService(request.user)
-        service.change_user_mode()
-
-        return Response(status=status.HTTP_200_OK)
