+++ users/views/view_v1_user.py
@@ -7,6 +7,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from users.serializers import UserOutputSerializer
+from users.services.service_v1_user import UserService
 
 
 class MyProfileAPI(APIView):
@@ -25,3 +26,18 @@ class MyProfileAPI(APIView):
         user = request.user
         output_serializer = UserOutputSerializer(user)
         return Response(output_serializer.data, status=status.HTTP_200_OK)
+
+
+class ChangeModeAPI(APIView):
+    @swagger_auto_schema(
+        operation_summary="V1 Change User Mode API",
+        operation_description="호스트-게스트 모드 변경",
+        responses={
+            status.HTTP_200_OK: openapi.Response("변경 완료"),
+        },
+    )
+    def patch(self, request: Request):
+        service = UserService(request.user)
+        service.change_user_mode()
+
+        return Response(status=status.HTTP_200_OK)
